requirements.txt
dj-database-url==0.5.0
Django==3.0.4
djangorestframework==3.11.0
gunicorn==20.0.4
pytz==2019.3
whitenoise==5.0.1
psycopg2-binary==2.8.4
Pillow==7.0.0
django-webpack-loader==0.7.0